+++ bin/cli.js
@@ -13,13 +13,14 @@ var argv = minimist(process.argv.slice(2), {
     'version'
   ],
   alias: {
-    h: 'help',
-    v: 'version',
+    b: 'config-basedir',
+    c: 'config',
     d: 'diff',
+    h: 'help',
+    i: 'ignore-path',
     l: 'list',
     R: 'recursive',
-    b: 'config-basedir',
-    c: 'config'
+    v: 'version',
   }
 })
 
@@ -36,12 +37,14 @@ if (argv.h) {
   console.log('')
   console.log('Options:')
   console.log('')
-  console.log('  -d, --diff             output diff against original file')
+  console.log('  -d, --diff             Output diff against original file')
   console.log('  -l, --list             Format list of space seperated files(globs) in place')
-  console.log('  -c, --config           path to a specific configuration file (JSON, YAML, or CommonJS)')
-  console.log('  -b, --config-basedir   path to the directory that relative paths defining "extends"')
-  console.log('  -v, --version          output the version number')
-  console.log('  -h, --help             output usage information')
+  console.log('  -c, --config           Path to a specific configuration file (JSON, YAML, or CommonJS)')
+  console.log('  -b, --config-basedir   Path to the directory that relative paths defining "extends"')
+  console.log('  -v, --version          Output the version number')
+  console.log('  -h, --help             Output usage information')
+  console.log('  -i, --ignore-path      Path to a file containing patterns that describe files to ignore.')
+  console.log('  --stdin-filename       A filename to assign stdin input.')
   process.exit()
 }
 
@@ -55,6 +58,10 @@ if (argv.b) {
   options.configBasedir = (path.isAbsolute(argv.b))
     ? argv.b
     : path.resolve(process.cwd(), argv.b)
+}
+
+if (argv.i) {
+  options.ignorePath = argv.i
 }
 
 if (argv.l) {
@@ -88,8 +95,12 @@ if (argv.l) {
     })
 } else {
   stdin(function (css) {
+    options.codeFilename = argv['stdin-filename']
     postcss([stylefmt(options)])
-      .process(css, { syntax: scss })
+      .process(css, {
+        from: options.codeFilename,
+        syntax: scss
+      })
       .then(function (result) {
         process.stdout.write(result.css)
       })