+++ index.js
@@ -10,17 +10,17 @@ const formatSassVariables = require('./lib/formatSassVariables')
 
 const stylefmt = postcss.plugin('stylefmt', function (options) {
   var paramer = params(options)
-  return function (root) {
-    return paramer(root).then(function (params) {
-      formatComments(root, params)
-      formatAtRules(root, params)
-      formatOrder(root, params)
-      formatRules(root, params)
-      formatSassVariables(root, params)
-      return root
+  return function (root, result) {
+    return paramer(root, result).then(function (params) {
+      if(params) {
+        formatComments(root, params)
+        formatAtRules(root, params)
+        formatOrder(root, params)
+        formatRules(root, params)
+        formatSassVariables(root, params)
+      }
     }).catch(function (err) {
       console.error(err.stack)
-      return root
     })
   }
 })